refactor(web): remove dead code + renomeia 4 *Redesign collision-free (038 Slice B1)
Deleta 7 bases pré-redesign mortas (0 importers, sem teste vivo):
InsightCard, SmartAlerts, StockCard, MedicineCard, ConsultationView,
ReminderSuggestion + PrescriptionTimeline (base+test+css — testava código morto).
Renomeia 4 *Redesign sem colisão → base (RingGauge, CostSummary,
PrescriptionTimeline, BottomNav): arquivo+css+identificador+importers.
Folder-dissolve + 6 renames com base morta ficam pro Slice B2.
