+++ CHANGELOG.md
@@ -83,6 +83,7 @@ Cuidando da sua rotina com carinho e zero complicação. 💙
 ### ♻️ Refatorado (estrutura web — aposenta naming "redesign", Slice A) — web · no-user-impact
 
 - **Spec 038 / Slice A:** o redesign da experiência do paciente foi entregue, mas a pasta de produção das views ainda se chamava `views/redesign/`. `git mv views/redesign/* → views/` (Dashboard, Stock, Treatments, Medicines, HealthHistory, Settings, Emergency, Profile, Consultation, NotificationInbox + subpastas `history/ profile/ settings/`); artefatos Landing agrupados em `views/landing/`. Atualizados `AppViewRouter` (12 lazy imports), `vite.config.js` (alias `@settings` + manualChunks `HealthHistory`/`Stock`/`Landing`), o resolver `@settings` do `eslint.config.js` (alias duplicado fora do vite — AP-238) e 3 arquivos de teste. Refator puro, sem mudança de comportamento. Slices B (dissolver `*Redesign`/dead code) e C (carve-out + doc) a seguir.
+- **Spec 038 / Slice B1:** removido dead code legado (7 componentes pré-redesign com 0 importers: `InsightCard`, `SmartAlerts`, `StockCard`, `MedicineCard`, `ConsultationView`, `ReminderSuggestion` + a base `PrescriptionTimeline` com test/css — testava código morto) e renomeados 4 componentes `*Redesign` collision-free para a base (`RingGaugeRedesign`→`RingGauge`, `CostSummaryRedesign`→`CostSummary`, `PrescriptionTimelineRedesign`→`PrescriptionTimeline`, `BottomNavRedesign`→`BottomNav` — arquivo + css + identificador + importers). −2843 linhas. Refator puro, sem mudança de comportamento. B2 (dissolver pastas `redesign/` + 6 renames com base morta) e C (carve-out + doc) a seguir.
 
 ### 🔒 Segurança / ✨ Melhorado (Chatbot IA — endurecimento) — web `4.10.0`→**`4.11.0`** · backend/infra · Telegram
