--- apps/web/src/features/dashboard/components/InsightCard.css
@@ -1,97 +1,0 @@
-/**
- * InsightCard Component Styles
- * 
- * Visual Example 3: Enhanced InsightCard with gradient and glassmorphism
- */
-
-.insight-card {
-  background: var(--gradient-insight);
-  border: var(--border-width-prominent) solid var(--glass-border-heavy);
-  border-radius: var(--radius-card-lg);
-  padding: var(--spacing-component-compact);
-  backdrop-filter: var(--glass-blur);
-  box-shadow: var(--shadow-card);
-  display: flex;
-  gap: var(--spacing-component-compact);
-  align-items: flex-start;
-  transition: var(--transition-hover);
-}
-
-.insight-card:hover {
-  box-shadow: var(--shadow-card-hover), var(--glow-hover);
-  transform: translateY(-2px);
-  border-color: var(--color-secondary);
-}
-
-.insight-card__icon {
-  font-size: var(--font-size-2xl);
-  flex-shrink: 0;
-  filter: drop-shadow(0 0 8px rgba(0, 240, 255, 0.4));
-}
-
-.insight-card__content {
-  flex: 1;
-  min-width: 0;
-}
-
-.insight-card__text {
-  font-size: var(--font-size-sm);
-  line-height: 1.5;
-  color: var(--text-primary-dark);
-}
-
-.insight-card__highlight {
-  color: var(--color-secondary);
-  font-weight: 600;
-  text-shadow: 0 0 8px rgba(0, 240, 255, 0.4);
-}
-
-.insight-card__action {
-  display: inline-flex;
-  align-items: center;
-  gap: var(--spacing-related-tight);
-  margin-top: var(--spacing-related);
-  padding: var(--spacing-component-tight) var(--spacing-component-normal);
-  background: var(--glass-bg-heavy);
-  border: var(--border-width-default) solid var(--glass-border);
-  border-radius: var(--radius-button);
-  color: var(--text-primary);
-  font-size: var(--font-size-xs);
-  font-weight: 600;
-  cursor: pointer;
-  transition: var(--transition-hover);
-}
-
-.insight-card__action:hover {
-  background: var(--color-secondary);
-  color: var(--text-on-primary);
-  border-color: var(--color-secondary);
-  box-shadow: var(--shadow-button), var(--glow-hover);
-}
-
-.insight-card__action:focus-visible {
-  outline: 2px solid var(--color-primary);
-  outline-offset: 2px;
-  transition: var(--transition-focus);
-}
-
-.insight-card__action:active {
-  transform: translateY(0) scale(0.98);
-  transition: var(--transition-active);
-}
-
-/* Responsive adjustments */
-@media (max-width: 480px) {
-  .insight-card {
-    flex-direction: column;
-  }
-
-  .insight-card__icon {
-    align-self: flex-start;
-  }
-
-  .insight-card__action {
-    width: 100%;
-    justify-content: center;
-  }
-}
