+++ apps/web/src/features/dashboard/components/RingGauge.css
