--- apps/web/src/features/dashboard/components/SmartAlerts.css
@@ -1,161 +1,0 @@
-.smart-alerts {
-  display: flex;
-  flex-direction: column;
-  gap: var(--spacing-list-compact); /* 8px - era 12px */
-  width: 100%;
-  margin-bottom: var(--spacing-section-minor); /* 12px - era 24px */
-}
-
-.smart-alert {
-  display: flex;
-  align-items: flex-start;
-  gap: var(--space-2); /* 8px */
-  padding: var(--space-3); /* 12px - mais compacto */
-  background: var(--glass-bg);
-  backdrop-filter: blur(4px); /* Mais sutil */
-  border: 1px solid var(--glass-border);
-  border-left: 3px solid var(--alert-color); /* 3px - era 4px */
-  border-radius: var(--radius-card-md); /* 12px */
-  box-shadow: var(--shadow-sm); /* Mais sutil */
-  animation: slide-in 0.3s ease-out;
-  transition: var(--transition-hover);
-}
-
-@keyframes slide-in {
-  from {
-    opacity: 0;
-    transform: translateY(-10px);
-  }
-  to {
-    opacity: 1;
-    transform: translateY(0);
-  }
-}
-
-.smart-alert:hover {
-  box-shadow: var(--shadow-card-hover);
-  transform: translateY(-1px); /* Mais sutil */
-}
-
-.smart-alert--critical {
-  --alert-color: var(--color-primary);
-  box-shadow: var(--shadow-alert-critical);
-}
-
-.smart-alert--warning {
-  --alert-color: var(--color-warning);
-}
-
-.smart-alert--info {
-  --alert-color: var(--color-secondary);
-}
-
-.smart-alert__icon {
-  font-size: var(--font-size-xl);
-  flex-shrink: 0;
-}
-
-.smart-alert__content {
-  flex: 1;
-  min-width: 0;
-}
-
-.smart-alert__title {
-  margin: 0;
-  font-size: var(--font-size-base);
-  font-weight: 700;
-  color: var(--text-primary-dark);
-  line-height: 1.3;
-}
-
-.smart-alert__message {
-  margin: var(--spacing-related-tight) 0 0 0;
-  font-size: var(--font-size-sm);
-  color: var(--text-secondary-dark);
-  line-height: 1.4;
-}
-
-.smart-alert__actions {
-  display: flex;
-  gap: var(--space-2); /* 8px */
-  margin-top: var(--space-2); /* 8px */
-}
-
-.smart-alert__btn {
-  padding: var(--space-1) var(--space-3); /* 4px 12px - mais compacto */
-  border-radius: var(--radius-button);
-  font-size: var(--font-size-xs); /* 12px */
-  font-weight: 600; /* Era 700 */
-  cursor: pointer;
-  transition: var(--transition-hover);
-  border: none;
-  text-transform: uppercase;
-  letter-spacing: 0.03em; /* Era 0.05em */
-  min-height: 28px; /* Touch target */
-}
-
-.smart-alert__btn:hover {
-  transform: translateY(-1px);
-}
-
-.smart-alert__btn:active {
-  transform: translateY(0) scale(0.98);
-  transition: var(--transition-active);
-}
-
-.smart-alert__btn:focus-visible {
-  outline: 2px solid var(--color-primary);
-  outline-offset: 2px;
-  transition: var(--transition-focus);
-}
-
-.smart-alert__btn--primary {
-  background: var(--alert-color);
-  color: var(--text-on-primary);
-  box-shadow: var(--shadow-button);
-}
-
-.smart-alert__btn--primary:hover {
-  box-shadow: var(--shadow-button-hover), var(--glow-hover);
-  transform: translateY(-1px);
-}
-
-.smart-alert__btn--secondary {
-  background: transparent;
-  border: 2px solid var(--color-secondary);
-  color: var(--color-secondary);
-  padding: var(--space-2) var(--space-4); /* 8px 16px - maior */
-  min-height: 36px; /* Touch target maior */
-  border-radius: var(--radius-button); /* Pill shape */
-  font-weight: 600;
-  text-transform: uppercase;
-  letter-spacing: 0.03em;
-  transition: var(--transition-hover);
-}
-
-.smart-alert__btn--secondary:hover {
-  background: rgba(6, 182, 212, 0.1); /* Cyan com 10% opacidade */
-  border-color: var(--color-secondary);
-  color: var(--color-secondary);
-  box-shadow: 0 0 12px rgba(6, 182, 212, 0.3); /* Glow sutil */
-  transform: translateY(-1px);
-}
-
-.smart-alert__btn--secondary:active {
-  transform: translateY(0) scale(0.98);
-}
-
-@media (max-width: 480px) {
-  .smart-alert {
-    flex-direction: column;
-  }
-
-  .smart-alert__actions {
-    width: 100%;
-    margin-top: var(--spacing-component-compact);
-  }
-
-  .smart-alert__btn {
-    flex: 1;
-  }
-}
