--- apps/web/src/features/medications/components/MedicineCard.css
@@ -1,99 +1,0 @@
-/* MedicineCard.css */
-.medicine-card {
-  display: flex;
-  flex-direction: column;
-  gap: var(--space-4);
-  cursor: default;
-  background: var(--glass-bg);
-  backdrop-filter: var(--glass-blur);
-  border: 1px solid var(--glass-border);
-  border-radius: var(--radius-lg);
-  padding: var(--space-5);
-  transition: var(--transition-hover);
-}
-
-.medicine-card:hover {
-  transform: scale(var(--scale-hover));
-  box-shadow: var(--glow-hover-primary);
-  border-color: var(--glass-border-heavy);
-}
-
-.medicine-header {
-  display: flex;
-  /*flex-direction: column;*/ /* Modified */
-  justify-content: space-between;
-  align-items: center; /* Added */
-  gap: var(--space-2);
-  padding-bottom: var(--space-3);
-  border-bottom: 1px solid var(--glass-border);
-}
-
-.medicine-name {
-  margin: 0;
-  font-size: var(--font-size-lg);
-  color: var(--text-primary);
-  font-weight: 600;
-}
-
-.medicine-ingredient {
-  font-size: var(--font-size-sm);
-  color: var(--color-primary);
-  font-weight: 500;
-}
-
-/* Dependency indicator styling */
-.dependency-indicator {
-  color: var(--color-warning);
-  font-size: 1.2em;
-  margin-left: var(--space-2); /* Add some spacing */
-  filter: drop-shadow(0 0 5px rgba(245, 158, 11, 0.5));
-}
-
-.medicine-details {
-  display: flex;
-  flex-direction: column;
-  gap: var(--space-3);
-}
-
-.detail-item {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  font-size: var(--font-size-sm);
-}
-
-.detail-label {
-  color: var(--text-secondary);
-  font-weight: 500;
-}
-
-.detail-value {
-  color: var(--text-primary);
-  font-weight: 600;
-}
-
-.medicine-actions {
-  display: flex;
-  gap: var(--space-3);
-  margin-top: var(--space-2);
-}
-
-.medicine-actions button {
-  flex: 1;
-  transition: var(--transition-hover);
-}
-
-.medicine-actions button:hover {
-  transform: scale(var(--scale-hover));
-  box-shadow: var(--glow-hover-primary);
-}
-
-.medicine-actions button:active {
-  transform: scale(var(--scale-active));
-}
-
-@media (max-width: 768px) {
-  .medicine-actions {
-    flex-direction: column;
-  }
-}
