+++ apps/web/src/features/stock/components/CostSummary.css
