--- apps/web/src/features/stock/components/PrescriptionTimelineRedesign.css
@@ -1,167 +1,0 @@
-.prescription-timeline-redesign {
-  background: var(--color-surface-container-lowest, #ffffff);
-  border-radius: var(--radius-card, 2rem);
-  padding: 1.75rem;
-  box-shadow: var(--shadow-ambient, 0 24px 24px rgba(25, 28, 29, 0.04));
-}
-
-.prescription-timeline-redesign__title {
-  font-family: var(--font-display, 'Public Sans', sans-serif);
-  font-weight: 700;
-  font-size: var(--text-headline-sm, 1.25rem);
-  color: var(--color-on-surface, #191c1d);
-  margin: 0 0 1.25rem;
-}
-
-.prescription-timeline-redesign__summary {
-  display: flex;
-  gap: 0.5rem;
-  flex-wrap: wrap;
-  margin-bottom: 1.25rem;
-}
-
-.prescription-timeline-redesign__list {
-  display: flex;
-  flex-direction: column;
-  gap: 1.25rem;
-}
-
-.prescription-timeline-redesign__item {
-}
-
-.prescription-timeline-redesign__item-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-bottom: 0.375rem;
-}
-
-.prescription-timeline-redesign__name {
-  font-weight: 600;
-  font-size: var(--text-body-sm, 0.875rem);
-  color: var(--color-on-surface, #191c1d);
-}
-
-.prescription-timeline-redesign__dates,
-.prescription-timeline-redesign__dates-row {
-  font-size: var(--text-label-md, 0.75rem);
-  color: var(--color-outline, #6d7a76);
-  font-variant-numeric: tabular-nums;
-  margin: 0 0 0.375rem;
-}
-
-.prescription-timeline-redesign__bar-track {
-  height: 8px;
-  background: var(--color-surface-container-low, #f2f4f5);
-  border-radius: var(--radius-full, 9999px);
-  overflow: hidden;
-}
-
-.prescription-timeline-redesign__bar-fill {
-  height: 100%;
-  border-radius: var(--radius-full, 9999px);
-  width: 0;
-}
-
-.prescription-timeline-redesign__bar-fill--ativa {
-  background: var(--color-primary, #006a5e);
-}
-
-.prescription-timeline-redesign__bar-fill--vencendo {
-  background: var(--color-tertiary, #7b5700);
-}
-
-.prescription-timeline-redesign__bar-fill--vencida {
-  background: var(--color-error, #ba1a1a);
-}
-
-.prescription-timeline-redesign__bar-fill--finalizada {
-  background: var(--color-outline, #6d7a76);
-}
-
-/* Timeline Fill animation */
-@media (prefers-reduced-motion: no-preference) {
-  .prescription-timeline-redesign__bar-fill {
-    animation: prxFill 1.2s cubic-bezier(0, 0, 0.2, 1) 0.3s both;
-  }
-
-  @keyframes prxFill {
-    from {
-      width: 0;
-    }
-    to {
-      width: var(--progress, 0%);
-    }
-  }
-}
-
-@media (prefers-reduced-motion: reduce) {
-  .prescription-timeline-redesign__bar-fill {
-    width: var(--progress, 0%);
-  }
-}
-
-/* Status badges reutilizados aqui */
-.prx-status {
-  display: inline-flex;
-  align-items: center;
-  font-size: var(--text-label-sm, 0.625rem);
-  font-weight: 600;
-  padding: 0.175rem 0.5rem;
-  border-radius: var(--radius-full, 9999px);
-}
-
-.prx-status--ativa {
-  background: color-mix(in srgb, var(--color-primary, #006a5e) 12%, transparent);
-  color: var(--color-primary, #006a5e);
-}
-
-.prx-status--vencendo {
-  background: color-mix(in srgb, var(--color-tertiary-fixed, #ffdea8) 60%, transparent);
-  color: var(--color-tertiary, #7b5700);
-}
-
-.prx-status--vencida {
-  background: color-mix(in srgb, var(--color-error, #ba1a1a) 12%, transparent);
-  color: var(--color-error, #ba1a1a);
-}
-
-.prx-status--finalizada {
-  background: color-mix(in srgb, var(--color-outline, #6d7a76) 12%, transparent);
-  color: var(--color-outline, #6d7a76);
-}
-
-/* ── Bar footer: datas + dias restantes ── */
-.prescription-timeline-redesign__bar-footer {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-top: 0.25rem;
-  gap: 0.5rem;
-}
-
-.prescription-timeline-redesign__dates-label {
-  font-size: var(--text-label-sm, 0.6875rem);
-  color: var(--color-outline, #6d7a76);
-  font-family: var(--font-body, Lexend, sans-serif);
-}
-
-.prescription-timeline-redesign__days-left {
-  font-size: var(--text-label-sm, 0.6875rem);
-  font-weight: 600;
-  font-family: var(--font-body, Lexend, sans-serif);
-  white-space: nowrap;
-}
-
-.prx-days--ativa {
-  color: var(--color-primary, #006a5e);
-}
-.prx-days--vencendo {
-  color: var(--color-tertiary, #7b5700);
-}
-.prx-days--vencida {
-  color: var(--color-error, #ba1a1a);
-}
-.prx-days--finalizada {
-  color: var(--color-outline, #6d7a76);
-}
