+++ apps/web/src/shared/components/ui/__tests__/BottomNav.test.jsx
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi, afterEach } from 'vitest'
 import { render, screen, fireEvent } from '@testing-library/react'
-import BottomNav from '@/shared/components/ui/BottomNavRedesign'
+import BottomNav from '@/shared/components/ui/BottomNav'
 
 describe('BottomNav', () => {
   afterEach(() => {
